finished mcu side, need to add more pins for more ADCs

+++ Controller/i2c_comm.h
@@ -3,6 +3,6 @@
 
 void I2C_received(uint8_t received_data); 
 void I2C_requested();
-extern uint32_t fData[NUMBER_OF_SENSORS];
+extern float fData[NUMBER_OF_SENSORS];
 volatile uint8_t data;
 char data_str[20*NUMBER_OF_SENSORS];
